Refactor ActorRuntime to return Result instead of storing errno
Remove hidden mutation from ActorRuntime trait by returning errno
directly from operations instead of storing it in last_errno field.

Changes:
- ActorRuntime methods now return Result<T, i32> where i32 is errno
- Remove get_errno() from trait (no longer needed)
- All trait methods use &self (no mutation)
- BlockingActorRuntime.mark_failed() takes Option<i32> explicitly
- Remove last_errno field from BlockingActorRuntime
- Update AReader/AWriter to use Result-based API
- errno_to_error_kind takes i32 (matches POSIX int)

This eliminates shared mutable state, making data flow explicit and
allowing AReader/AWriter to hold shared &dyn ActorRuntime references.

+++ ailets-rs/actor_io/src/error_mapping.rs
@@ -6,7 +6,7 @@
 /// Convert errno to `embedded_io::ErrorKind`
 #[must_use]
 #[allow(clippy::match_same_arms)] // We explicitly list common errno values for documentation
-pub fn errno_to_error_kind(errno: isize) -> embedded_io::ErrorKind {
+pub fn errno_to_error_kind(errno: i32) -> embedded_io::ErrorKind {
     match errno {
         1 | 13 => embedded_io::ErrorKind::PermissionDenied, // EPERM, EACCES
         2 => embedded_io::ErrorKind::NotFound,              // ENOENT
